requirements.txt
sentence-transformers
ctransformers
langchain
langchain-chroma
pypdf2
faiss-cpu
python-dotenv